requirements.txt
streamlit
langchain
langchain-google-genai
google-generativeai
PyPDF2
python-dotenv
faiss-cpu
pandas
